// Qanat.Web/src/app/shared/generated/model/parcel-changes-grid-item-dto.ts
export class ParcelChangesGridItemDto { 
    ParcelID?: number;
    ParcelNumber?: string | null;
    GeographyID?: number;
    WaterAccount?: WaterAccountLinkDisplayDto;
    ParcelStatus?: ParcelStatusSimpleDto;
    IsReviewed?: boolean;
    ReviewDate?: string | null;
    CurrentParcelHistoryUploadDate?: string;
    PreviousParcelHistoryUploadDate?: string | null;
    ParcelFieldDiffs?: Array<ParcelFieldDiffDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}